Add actions to Simple K8s test
This commit changes the package used for the test
k8s-03. The new package has actions included and
uses local k8s charms.

This required the additon of the upload-package
commands in the robot test libraries and the
K8s Ns Action command.

Additionally the timeout time for the k8scluster
creation has been extended, as it was to short for
some test environments.

Change-Id: I846733982a7874b3467270d1a1d1fa8a330030f8
Signed-off-by: Dominik Fleischmann <dominik.fleischmann@canonical.com>